refactor: share the no-cache header helper across adapters and drop the per-result wrappers
Move SetNoCacheHeaders, and its pre-computed Cache-Control value, into the shared
Abblix.Oidc.Server.AspNetCore project so the MVC and Minimal API adapters keep a
single copy instead of byte-identical duplicates. The MVC GeneratedHtmlResult and the
Minimal API endpoint-level no-cache filter both call the shared helper.

Drop the redundant per-result WithNoCacheHeaders from the MVC token and end-session
formatters: the controllers already carry [ResponseCache(NoStore)], so wrapping each
ActionResult only duplicated the header. The now-unused MVC WithNoCacheHeaders helper
is removed with it.
