+++ Abblix.Oidc.Server.AspNetCore/HttpResponseExtensions.cs
@@ -1,0 +1,69 @@
+// Abblix OIDC Server Library
+// Copyright (c) Abblix LLP. All rights reserved.
+//
+// DISCLAIMER: This software is provided 'as-is', without any express or implied
+// warranty. Use at your own risk. Abblix LLP is not liable for any damages
+// arising from the use of this software.
+//
+// LICENSE RESTRICTIONS: This code may not be modified, copied, or redistributed
+// in any form outside of the official GitHub repository at:
+// https://github.com/Abblix/OIDC.Server. All development and modifications
+// must occur within the official repository and are managed solely by Abblix LLP.
+//
+// Unauthorized use, modification, or distribution of this software is strictly
+// prohibited and may be subject to legal action.
+//
+// For full licensing terms, please visit:
+//
+// https://oidc.abblix.com/license
+//
+// CONTACT: For license inquiries or permissions, contact Abblix LLP at
+// info@abblix.com
+
+using Microsoft.AspNetCore.Http;
+using Microsoft.Net.Http.Headers;
+
+namespace Abblix.Oidc.Server.AspNetCore;
+
+/// <summary>
+/// Response-shaping helpers shared by the MVC and Minimal API adapters.
+/// </summary>
+public static class HttpResponseExtensions
+{
+    /// <summary>
+    /// Pre-computed Cache-Control header value that combines multiple cache prevention directives
+    /// for maximum compatibility across browsers, proxies, and HTTP versions.
+    /// </summary>
+    private static readonly CacheControlHeaderValue PreventStorageInAnyCache = new()
+    {
+        MaxAge = TimeSpan.Zero,
+        SharedMaxAge = TimeSpan.Zero,
+        NoStore = true,
+        NoCache = true,
+    };
+
+    /// <summary>
+    /// Sets comprehensive no-cache headers on the response to prevent caching.
+    /// Ensures responses containing sensitive information (tokens, credentials, logout pages) are never cached.
+    /// </summary>
+    /// <remarks>
+    /// Sets the following headers for maximum compatibility:
+    /// <list type="bullet">
+    /// <item><description>
+    /// <b>Cache-Control</b>: "no-store, no-cache, max-age=0, s-maxage=0"
+    /// - Prevents storage in any cache (HTTP/1.1)</description></item>
+    /// <item><description>
+    /// <b>Pragma</b>: "no-cache" - Prevents caching in HTTP/1.0 proxies and legacy browsers</description></item>
+    /// <item><description>
+    /// <b>Expires</b>: Unix epoch - Sets expiration to the past for HTTP/1.0 compatibility</description></item>
+    /// </list>
+    /// </remarks>
+    /// <param name="response">The HTTP response to modify.</param>
+    public static void SetNoCacheHeaders(this HttpResponse response)
+    {
+        var headers = response.GetTypedHeaders();
+        headers.Expires = DateTimeOffset.UnixEpoch;
+        headers.CacheControl = PreventStorageInAnyCache;
+        response.Headers.Pragma = CacheControlHeaderValue.NoCacheString;
+    }
+}
