+++ Abblix.Oidc.Server.MinimalApi/EndpointRouteBuilderExtensions.cs
@@ -22,6 +22,7 @@
 
 using System.Net.Http.Headers;
 using Abblix.Jwt;
+using Abblix.Oidc.Server.AspNetCore;
 using Abblix.Oidc.Server.Common.Configuration;
 using Abblix.Oidc.Server.Common.Constants;
 using Abblix.Oidc.Server.Common.Interfaces;
