+++ Abblix.Oidc.Server.MinimalApi/Formatters/RegistrationClientUriBuilder.cs
