+++ Abblix.Oidc.Server.Mvc/ActionResults/ActionResultExtensions.cs
@@ -87,51 +87,6 @@ public static class ActionResultExtensions
 			});
 
 	/// <summary>
-	/// Pre-computed Cache-Control header value that combines multiple cache prevention directives
-	/// for maximum compatibility across browsers, proxies, and HTTP versions.
-	/// </summary>
-	private static readonly CacheControlHeaderValue PreventStorageInAnyCache = new()
-	{
-		MaxAge = TimeSpan.Zero,
-		SharedMaxAge = TimeSpan.Zero,
-		NoStore = true,
-		NoCache = true,
-	};
-
-	/// <summary>
-	/// Sets comprehensive no-cache headers on the response to prevent caching.
-	/// Ensures responses containing sensitive information (tokens, credentials, logout pages) are never cached.
-	/// </summary>
-	/// <remarks>
-	/// Sets the following headers for maximum compatibility:
-	/// <list type="bullet">
-	/// <item><description>
-	/// <b>Cache-Control</b>: "no-store, no-cache, max-age=0, s-maxage=0"
-	/// - Prevents storage in any cache (HTTP/1.1)</description></item>
-	/// <item><description>
-	/// <b>Pragma</b>: "no-cache" - Prevents caching in HTTP/1.0 proxies and legacy browsers</description></item>
-	/// <item><description>
-	/// <b>Expires</b>: Unix epoch - Sets expiration to the past for HTTP/1.0 compatibility</description></item>
-	/// </list>
-	/// </remarks>
-	/// <param name="response">The HTTP response to modify.</param>
-	public static void SetNoCacheHeaders(this HttpResponse response)
-	{
-		var headers = response.GetTypedHeaders();
-		headers.Expires = DateTimeOffset.UnixEpoch;
-		headers.CacheControl = PreventStorageInAnyCache;
-		response.Headers.Pragma = CacheControlHeaderValue.NoCacheString;
-	}
-
-	/// <summary>
-	/// Decorates an <see cref="ActionResult"/> to prevent caching by appending comprehensive no-cache headers.
-	/// </summary>
-	/// <param name="innerResult">The <see cref="ActionResult"/> to decorate.</param>
-	/// <returns>A decorated <see cref="ActionResult"/> with comprehensive cache prevention headers.</returns>
-	public static ActionResult WithNoCacheHeaders(this ActionResult innerResult)
-		=> new ActionResultDecorator(innerResult, SetNoCacheHeaders);
-
-	/// <summary>
 	/// Formats an <see cref="OidcError"/> as an appropriate HTTP error response per RFC 6750 Section 3.
 	/// Bearer token errors (<c>invalid_token</c>) return HTTP 401 with only a <c>WWW-Authenticate</c> header
 	/// and no response body. Scope errors (<c>insufficient_scope</c>) return HTTP 403 with the header.
