+++ Abblix.Oidc.Server.Mvc/ActionResults/GeneratedHtmlResult.cs
@@ -20,6 +20,7 @@
 // CONTACT: For license inquiries or permissions, contact Abblix LLP at
 // info@abblix.com
 
+using Abblix.Oidc.Server.AspNetCore;
 using System.Net.Mime;
 using System.Xml;
 using Microsoft.AspNetCore.Http;
